+++ src/meson.build
@@ -33,9 +33,8 @@ timeswitch_sources = [
   'main.py',
   'main_window.py',
   'main_window_shortcuts.py',
-  'manage_presets_window.py',
+  'manage_presets_view.py',
   'player.py',
-  'presets_manager.py',
   'timer.py'
 ]
